refactor: join 패키지명 auth로 변경
