+++ src/main/java/com/vata/auth/dto/LoginRequest.java
@@ -1,0 +1,3 @@
+package com.vata.auth.dto;
+
+public record LoginRequest (String email, String password) {}
