--- src/main/java/com/vata/join/dto/LoginRequest.java
@@ -1,3 +1,0 @@
-package com.vata.join.dto;
-
-public record LoginRequest (String email, String password) {}
